==> notebook/01_data_extraction_bigquery.ipynb <==
%load_ext bigquery_magics

%%bigquery
SELECT *
FROM `bigquery-public-data.samples.gsod`
LIMIT 5


%%bigquery df_task1
SELECT
  station_number,
  wban_number,
  year,
  month,
  day,
  FORMAT_DATE('%Y-%m-%d', DATE(year, month, day)) AS date_ymd,
  mean_temp,
  min_temperature,
  max_temperature,
  total_precipitation,
  snow_depth,
  snow,
  rain,
  fog,
  hail,
  thunder,
  tornado
FROM `bigquery-public-data.samples.gsod`
WHERE year BETWEEN 2000 AND 2005
  AND station_number BETWEEN 725300 AND 726300


import pandas as pd
import os
from pathlib import Path

out_path =  Path("../data/interim/part1_task1.csv")


display(df_task1.head())
df_task1.info()
print("Rows, Columns:", df_task1.shape)

# To ensure date_ymd is string
df_task1['date_ymd'] = df_task1['date_ymd'].astype(str)
df_task1['date'] = pd.to_datetime(df_task1['date_ymd'], errors='coerce')

# save an interim copy
df_task1.to_csv(out_path, index=False)


%%bigquery year_counts
SELECT year, COUNT(*) AS n
FROM `bigquery-public-data.samples.gsod`
WHERE year BETWEEN 2000 AND 2005
  AND station_number BETWEEN 725300 AND 726300
GROUP BY year
ORDER BY year

# select the data between 2000-2005 for station numbers including and between 725300 and 726300



year_counts

df_task1.isna().mean().sort_values(ascending=False).head(10)
